requirements.txt
contourpy==1.3.2
cycler==0.12.1
-e git+ssh://git@github.com/yjclegend/distortion.git@75e95cb759a4b236f9480804baa5d96c7380b36a#egg=distortion
fonttools==4.59.2
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
